requirements.txt
customtkinter~=5.2.2
pyperclip~=1.9.0
pillow~=10.4.0
notify_py~=0.3.43
